refactor(telescope): optimize code structure and add authorization
- Move IncomingEntry to Storage namespace for better organization
- Remove unused BroadcastRecordingPipeMessage and EntryQueryOptions classes
- Add new Authorize middleware for telescope access control
- Remove redundant watcher() method from controllers
- Update related import statements and configurations
