+++ src/telescope/src/Controller/QueriesController.php
@@ -32,13 +32,4 @@ class QueriesController extends EntryController
     {
         return 'query';
     }
-
-    /**
-     * The watcher class for the controller.
-     */
-    protected function watcher()
-    {
-        // return RequestWatcher::class;
-        return null;
-    }
 }
