+++ src/telescope/src/Middleware/Authorize.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of friendsofhyperf/components.
+ *
+ * @link     https://github.com/friendsofhyperf/components
+ * @document https://github.com/friendsofhyperf/components/blob/main/README.md
+ * @contact  huangdijia@gmail.com
+ */
+
+namespace FriendsOfHyperf\Telescope\Middleware;
+
+use FriendsOfHyperf\Telescope\TelescopeConfig;
+use Hyperf\HttpServer\Contract\ResponseInterface as HttpResponse;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+
+class Authorize implements MiddlewareInterface
+{
+    public function __construct(
+        protected TelescopeConfig $telescopeConfig,
+        protected HttpResponse $response
+    ) {
+    }
+
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+    {
+        return $this->telescopeConfig->isEnabled() ? $handler->handle($request) : $this->response->html('')->withStatus(403);
+    }
+}
