+++ src/telescope/src/TelescopeContext.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
 
 namespace FriendsOfHyperf\Telescope;
 
+use FriendsOfHyperf\Telescope\Storage\IncomingEntry;
 use Hyperf\Context\Context;
 use Hyperf\Contract\PackerInterface;
 use Hyperf\Stringable\Str;
